feat()Adding carload endpoint

+++ src/app/driver/driver.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Driver } from '../models/driver';
 import { DriverService } from '../services/driver.service';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import {NzModalService} from 'ng-zorro-antd/modal';
 
 @Component({
   selector: 'app-driver',
@@ -17,18 +18,19 @@ export class DriverComponent implements OnInit {
   driverForm!: FormGroup;
 
   constructor(private driverService: DriverService,
-              private fb: FormBuilder) {
+              private fb: FormBuilder,
+              private modal: NzModalService) {
     this.initForm();
   }
 
   ngOnInit(): void {
     this.loadDrivers();
+
   }
 
   private loadDrivers(): void {
     this.driverService.getDrivers().subscribe(drivers => {
       this.listOfDisplayData = drivers;
-      console.log(drivers);
       this.totalDrivers = drivers.length;
     });
   }
@@ -39,23 +41,33 @@ export class DriverComponent implements OnInit {
 
   closeDriverDrawer(): void {
     this.isDriverDrawerVisible = false;
-    this.driverForm.reset({ status: 'ATIVO' });
+    this.driverForm.reset({ status: 'ACTIVO' });
   }
 
   submitDriver(): void {
     if (this.driverForm.valid) {
       this.driverService.addDriver(this.driverForm.value).subscribe(() => {
         this.loadDrivers();
+
         this.closeDriverDrawer();
       });
     }
   }
 
   deleteDriver(driver: Driver): void {
-    this.driverService.deleteDriver(driver.id).subscribe(() => {
-      this.loadDrivers();
+    this.modal.confirm({
+      nzTitle: 'Tens certeza que quer eliminar o Driver?',
+      nzContent: `Motorista: <strong>${driver.Name}</strong>`,
+      nzOkText: 'Sim',
+      nzOkType: 'primary',
+      nzCancelText: 'Não',
+      nzOnOk: () =>
+        this.driverService.deleteDriver(driver.id).subscribe(() => {
+          this.loadDrivers();
+        })
     });
   }
+
 
   search(): void {
     // Implementa aqui o filtro de pesquisa
@@ -66,7 +78,11 @@ export class DriverComponent implements OnInit {
       name: ['', Validators.required],
       phone: ['', [Validators.required, Validators.pattern('^[0-9]+$')]],
       carDescription: ['', Validators.required],
-      status: ['ATIVO', Validators.required]
+      status: ['ACTIVO', Validators.required]
     });
+  }
+
+  editDriver(driver: Driver) {
+
   }
 }
